implement complete methods in cell calendar page
