+++ lib/cell_calendar.dart
@@ -2,4 +2,5 @@ library cell_calendar;
 
 export 'src/calendar_event.dart';
 export 'src/cell_calendar.dart';
+export 'src/controllers/cell_calendar_page_controller.dart';
 export 'src/date_extension.dart';
